Workaround for a non-standard vxi11 implementation of the RPC Reply to a GET PORT Call (PortMapper).
Workaround for some equipment which do not set the last fragment bit to 1 for a RPC Reply to a GET PORT Call.
Therefore the last fragment is asserted True only when the length of the received data matches the announced length, and the length of the payload corresponds to the length of such a Reply to GET PORT Call (28).
The issue is described in https://groups.google.com/g/python-ivi/c/vxrdshKG43E/m/plOXFbsopUgJ?pli=1
The proposed workaround has been succesfully tested with equipment from Anritsu, namely the OSA models MS9740A and MS9740B.

Signed-off-by: Florian FRANK <florian1.frank@orange.com>